Refactor remote.py to remote subpackage
