+++ ai_models/remote/__init__.py
@@ -1,0 +1,4 @@
+from .api import RemoteAPI
+from .model import RemoteModel
+
+__all__ = ["RemoteAPI", "RemoteModel"]
